Rename: 컴포넌트 쓰임에 맞게 이름 변경
