+++ src/components/ConfirmationModal/ConfirmationModal.jsx
@@ -14,8 +14,8 @@ export default function ConfirmationModal({
   confirmNotificationMessage = '확인되었습니다.',
   cancelNotificationMessage,
   buttonStyles = {
-    cancel: 'rounded bg-gray200 px-4 py-2 hover:bg-gray300',
-    confirm: 'rounded bg-primary px-4 py-2 text-white hover:bg-indigo-700',
+    cancel: 'rounded bg-gray200 px-6 py-2 hover:bg-gray300',
+    confirm: 'rounded bg-primary px-6 py-2 text-white hover:bg-indigo-700',
   },
 }) {
   const dialogRef = useRef(null);
@@ -73,11 +73,11 @@ export default function ConfirmationModal({
     <>
       <dialog ref={dialogRef} className={`fixed inset-0 z-50 overflow-y-auto`}>
         <div className={`fixed inset-0 bg-black bg-opacity-50`}>
-          <div className="flex items-center justify-center min-h-screen">
-            <div className="p-6 bg-white rounded-lg shadow-lg">
+          <div className="flex min-h-screen items-center justify-center">
+            <div className="rounded-lg bg-white p-6 shadow-lg">
               <h2 className="mb-4 text-lg font-bold">{title}</h2>
               <p>{message}</p>
-              <div className="flex justify-end gap-3 mt-6">
+              <div className="mt-6 flex justify-end gap-3">
                 {showCancelButton && (
                   <button
                     ref={cancelButtonRef}
@@ -112,7 +112,7 @@ export default function ConfirmationModal({
 
       {notification && (
         <div className="fixed inset-0 flex items-center justify-center">
-          <div className="px-4 py-2 text-white bg-gray-800 bg-opacity-75 rounded">
+          <div className="rounded bg-gray-800 bg-opacity-75 px-4 py-2 text-white">
             {notification}
           </div>
         </div>
